microservice/requirements.txt

black==22.8.0
bleach==5.0.1
boto==2.49.0
boto3==1.16.63
botocore==1.19.63
boxing==0.1.4
cachetools==5.3.0
certifi==2021.10.8
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
cycler==0.11.0
Cython==3.0.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
Django==4.1.3
django-annoying==0.10.6
django-cors-headers==3.6.0
django-debug-toolbar==3.2.1
django-extensions==3.1.0
django-filter==2.4.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rest-swagger==2.2.0
django-rq==2.5.1
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.13.1
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
drf-yasg==1.20.0
expiringdict==1.2.2
filelock==3.12.2
fonttools==4.42.0
google-api-core==2.10.0
google-auth==2.11.0
google-cloud-appengine-logging==1.1.0
google-cloud-audit-log==0.2.0
google-cloud-core==2.3.2
google-cloud-logging==2.7.2
google-cloud-storage==2.5.0
google-crc32c==1.5.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.3
grpcio==1.54.0
grpcio-status==1.48.2
htmlmin==0.1.12
importlib-resources==6.0.1
imutils==0.5.4
inflection==0.5.1
isodate==0.6.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.2.0
jsonschema==3.2.0
kiwisolver==1.4.4
launchdarkly-server-sdk==7.5.0
lazy-object-proxy==1.7.1
lazy_loader==0.3
llvmlite==0.39.1
lockfile==0.12.2
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib
mccabe==0.7.0
mpmath==1.3.0
msgpack==1.0.4
mypy==0.971
mypy-extensions==0.4.3
networkx==3.1
ninja==1.11.1
nltk==3.6.7
numpy==1.25.2
openapi-codec==1.3.2
opencv-contrib-python-headless
ordered-set==4.0.2
packaging==23.0
pandas==1.3.5
pathspec==0.10.1
Pillow==9.5.0
platformdirs==2.5.2
pooch==1.7.0
proto-plus==1.22.2
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.6
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post4
pydantic==1.10.7
pygame==2.1.2
pylint==2.15.3
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.19.3
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.1
python-json-logger==2.0.4
pytz==2019.3
PyWavelets==1.4.1
PyYAML==6.0
redis==4.5.5
regex==2023.5.5
rq==1.10.1
rsa==4.9
ruamel.yaml==0.17.26
ruamel.yaml.clib==0.2.7
rules==2.2
s3transfer==0.3.7
scikit-image==0.21.0
scikit-learn==1.2.1
scipy==1.10.1
seaborn==0.12.2
semver==2.13.0
sentry-sdk==1.22.2
shapely==2.0.1
simplejson==3.19.1
soundfile==0.12.1
soxr==0.3.4
sqlparse==0.4.3
supervision==0.13.0
sympy==1.12
tesseract==0.1.3
threadpoolctl==3.1.0
tifffile==2023.8.12
tomli==2.0.1
tomlkit==0.11.4
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.0

ultralytics==8.0.150
uritemplate==4.1.1
urllib3==1.26.15
user-agents==2.2.0
webencodings==0.5.1
wrapt==1.14.1
xmljson==0.2.0
zipp==3.16.2

pip~=22.0.2
toml~=0.10.2
wheel~=0.37.1
cryptography~=41.0.3
py~=1.11.0
anyio~=3.7.1
sniffio~=1.3.0
pytest~=6.0.1
fastapi~=0.100.1
starlette~=0.27.0
requests~=2.28.0
websockets~=11.0.3
setuptools~=58.1.0
gunicorn
uvicorn